backEnd/requirements.txt
Django
djangorestframework
pytz
django-cleanup
requests
gunicorn
Pillow
django-cors-headers
faiss-cpu
googlemaps
git+https://github.com/facebookresearch/fastText.git
numpy
dj-static
